Aperette/requirements.txt
djangorestframework
Pillow
pymysql
django-cors-headers